recognition/position/april/requirements.txt
pyapriltags
numpy
opencv-python
requests
toml
colorama
dataclasses
Flask
pyzmq
keyboard